create the seccessfuly project

+++ src/component/signup/User.jsx
@@ -1,6 +1,8 @@
 
+
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 const User = () => {
@@ -11,18 +13,39 @@ const User = () => {
   const [users, setUsers] = useState(loadedUsers);
 
   const handleDelete = id => {
-    fetch(`http://localhost:5000/user/${id}`, {
-      method: 'DELETE'
+    console.log(id);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
     })
-    .then(res => res.json())
-      .then(data => {
-        if (data.deletedCount > 0) {
-          console.log('deleted successfuly');
-          
-          const remainingUsers = users.filter(user => user._id !== id);
-          setUsers(remainingUsers);
+      .then(result => {
+        if (result.isConfirmed) {
+        fetch(`http://localhost:5000/user/${id}`, {
+          method: 'DELETE',
+        })
+          .then(res => res.json())
+          .then(data => {
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: 'Deleted!',
+                text: 'Your cocacola has been deleted.',
+                icon: 'success',
+              });
+
+              const remainingUsers = users.filter(user => user._id !== id);
+              setUsers(remainingUsers);
+            }
+          });
       }
     })
+
+    
+    
   }
 
   return (
